requirements.txt
langchain-core
langchain_community 
langgraph 
langchain-openai
langchain_chroma
chromadb>=1.0.20,<2.0.0
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
sentence-transformers
torch
openpyxl
python-multipart